air_chatbot/requirements.txt
google-generativeai
opencv-python
numpy
Pillow
python-dotenv>=1.0.0
websockets
flask
flask-cors
openai>=1.12.0
requests
langchain>=0.2.17
langchain-community>=0.2.19
langchain-core>=0.2.43
langgraph>=0.3.5
langchain-openai>=0.1.25
langchain-text-splitters>=0.2.4
SQLAlchemy>=2.0.27
aiosqlite>=0.19.0
pydantic>=2.7.0
pydantic-settings>=2.8.1
langchain-groq
langgraph-supervisor
scikit-image>=0.18.0
scipy>=1.6.0 
edge-tts
pygame
pyttsx3
gtts
transformers 
torch
deep-translator
langdetect
selenium
pyautogui
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pygame